+++ transactionauthorizer/app/transactionauthorizer-usecases/src/main/kotlin/br/laiza/transactionauthorizer/usecases/dto/TransactionRequest.kt
@@ -1,8 +1,10 @@
 package br.laiza.transactionauthorizer.usecases.dto
 
+import java.math.BigDecimal
+
 class TransactionRequest(
     val account: String,
     var mcc: String,
-    val totalAmount: Double,
+    val totalAmount: BigDecimal,
     val merchant: String,
 )
