+++ txauthorizerconsumer/app/txauthorizerconsumer-core/src/main/kotlin/br/laiza/txauthorizerconsumer/core/entities/Transaction.kt
@@ -2,6 +2,7 @@ package br.laiza.txauthorizerconsumer.core.entities
 
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Entity
@@ -17,7 +18,7 @@ data class Transaction(
     val account: Account,
 
     @Column(nullable = false)
-    var totalAmount: Double,
+    var totalAmount: BigDecimal,
 
     @Column(nullable = false)
     var merchant: String,
