+++ txauthorizerconsumer/app/txauthorizerconsumer-core/src/main/kotlin/br/laiza/txauthorizerconsumer/core/message/TransactionMessage.kt
@@ -1,13 +1,16 @@
 package br.laiza.txauthorizerconsumer.core.message
+
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class TransactionMessage(
     val account: String,
     val transactionDate: LocalDateTime,
-    val amountTransaction: Double,
-    val newAmountWalletFood: Double,
-    val newAmountWalletMeal: Double,
-    val newAmountWalletCash: Double,
+    val amountTransaction: BigDecimal,
+    val newAmountWalletFood: BigDecimal,
+    val newAmountWalletMeal: BigDecimal,
+    val newAmountWalletCash: BigDecimal,
     val mcc: String,
     val merchant: String,
 )
+
